[autothrottle] separating legacy throttle code
